requirements.txt
aiocsv==1.2.1
aiofiles==0.5.0
aniso8601==7.0.0
anyio==3.3.2
async-exit-stack==1.0.1
async-generator==1.10
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
databases==0.5.2
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.68.1
geographiclib==1.52
geopy==2.2.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.1
h11==0.12.0
httpcore==0.13.7
httptools==0.1.2
httpx==0.19.0
idna==3.2
iniconfig==1.1.1
ipaddress==1.0.23
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
motor==2.5.1
nominatim==0.1
orjson==3.6.3
packaging==21.0
pluggy==1.0.0
promise==2.3
py==1.10.0
pydantic==1.8.2
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
rfc3986==1.5.0
Rx==1.6.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.25
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.13.4
uvloop==0.16.0
watchgod==0.7
websockets==8.1
